feat(consensus-any): move AnyTxType, UnknownTxEnvelope, UnknownTypedTransaction from alloy-network
Partially resolves #1598.

Move the three 'unknown' transaction types from alloy-network into
alloy-consensus-any, reducing the dependency surface for crates that
only need to work with unknown transaction envelopes:

- AnyTxType
- UnknownTypedTransaction
- UnknownTxEnvelope (+ internal DeserMemo helper)

Key changes:
- New crates/consensus-any/src/unknown.rs with the moved types
- alloy-consensus-any Cargo.toml: enable serde + alloy-serde when std
  feature is active (required by OtherFields and OnceLock)
- alloy-network re-exports the moved types from alloy-consensus-any
  for backwards compatibility
- AnyTypedTransaction and AnyTxEnvelope remain in alloy-network; moving
  them requires resolving the From<Self::UnsignedTx/TxEnvelope> bounds
  on the Network::TransactionRequest associated type, which would create
  a circular dependency between alloy-consensus-any and alloy-rpc-types-eth
